requirements.txt
langchain
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
langchain-openai
langchain_core
dill
pypdf
InstructorEmbedding
sentence-transformers
chromadb